requirements.txt
fastapi
uvicorn
python-dotenv
openai
httpx
requests
sqlmodel
postgrest-py

reflex==0.7.5
